apps/api-python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pyjwt>=2.8.0
httpx>=0.27.0
pydantic>=2.7.0
python-multipart>=0.0.9
